+++ src/pauxil/HPL_dlaswp01N.c
@@ -1,6 +1,6 @@
 /* 
  * -- High Performance Computing Linpack Benchmark (HPL)                
- *    HPL - 1.0a - January 20, 2004                          
+ *    HPL - 1.0b - December 15, 2004                          
  *    Antoine P. Petitet                                                
  *    University of Tennessee, Knoxville                                
  *    Innovative Computing Laboratories                                 
@@ -159,9 +159,9 @@ void HPL_dlaswp01N
    {
       for( i = 0; i < M; i++ )
       {
-         a0 = A + LINDXA[i];
-         if( LINDXAU[i] >= 0 ) { a1 = U + LINDXAU[i]; lda1 = LDU; }
-         else                  { a1 = A - LINDXAU[i]; lda1 = LDA; }
+         a0 = A + (size_t)(LINDXA[i]);
+         if( LINDXAU[i] >= 0 ) { a1 = U + (size_t)(LINDXAU[i]); lda1 = LDU; }
+         else                  { a1 = A - (size_t)(LINDXAU[i]); lda1 = LDA; }
 
          *a1 = *a0; a1 += lda1; a0 += LDA;
 #if ( HPL_LASWP01N_DEPTH >  1 )
@@ -197,10 +197,9 @@ void HPL_dlaswp01N
    {
       for( i = 0; i < M; i++ )
       {
-         a0 = A + LINDXA[i];
-         if( LINDXAU[i] >= 0 ) { a1 = U + LINDXAU[i]; lda1 = LDU; }
-         else                  { a1 = A - LINDXAU[i]; lda1 = LDA; }
-
+         a0 = A + (size_t)(LINDXA[i]);
+         if( LINDXAU[i] >= 0 ) { a1 = U + (size_t)(LINDXAU[i]); lda1 = LDU; }
+         else                  { a1 = A - (size_t)(LINDXAU[i]); lda1 = LDA; }
          for( j = 0; j < nr; j++, a1 += lda1, a0 += LDA ) { *a1 = *a0; }
       }
    }
