+++ src/pauxil/HPL_dlaswp03N.c
@@ -1,6 +1,6 @@
 /* 
  * -- High Performance Computing Linpack Benchmark (HPL)                
- *    HPL - 1.0a - January 20, 2004                          
+ *    HPL - 1.0b - December 15, 2004                          
  *    Antoine P. Petitet                                                
  *    University of Tennessee, Knoxville                                
  *    Innovative Computing Laboratories                                 
@@ -146,7 +146,8 @@ void HPL_dlaswp03N
    {
       for( i = 0; i < M; i++ )
       {
-         u0 = U + (int)(*(W0+i*LDW)); w0 = w + i * LDW;
+         u0 = U + (size_t)(*( W0 + (size_t)(i) * (size_t)(LDW) )); 
+         w0 = w + (size_t)(i) * (size_t)(LDW);
 
          *u0 = w0[ 0]; u0 += LDU;
 #if ( HPL_LASWP03N_DEPTH >  1 )
@@ -182,7 +183,8 @@ void HPL_dlaswp03N
    {
       for( i = 0; i < M; i++ )
       {
-         u0 = U + (int)(*(W0+i*LDW)); w0 = w + i * LDW;
+         u0 = U + (size_t)(*( W0 + (size_t)(i) * (size_t)(LDW) )); 
+         w0 = w + (size_t)(i) * (size_t)(LDW);
          for( j = 0; j < nr; j++, u0 += LDU ) { *u0 = w0[j]; }
       }
    }
