+++ src/pauxil/HPL_dlaswp05N.c
@@ -1,6 +1,6 @@
 /* 
  * -- High Performance Computing Linpack Benchmark (HPL)                
- *    HPL - 1.0a - January 20, 2004                          
+ *    HPL - 1.0b - December 15, 2004                          
  *    Antoine P. Petitet                                                
  *    University of Tennessee, Knoxville                                
  *    Innovative Computing Laboratories                                 
@@ -149,7 +149,7 @@ void HPL_dlaswp05N
    {
       for( i = 0; i < M; i++ )
       {
-         a0 = A + LINDXA[i]; u0 = U0 + LINDXAU[i];
+         a0 = A + (size_t)(LINDXA[i]); u0 = U0 + (size_t)(LINDXAU[i]);
 
          *a0 = *u0; a0 += LDA; u0 += LDU;
 #if ( HPL_LASWP05N_DEPTH >  1 )
@@ -185,7 +185,7 @@ void HPL_dlaswp05N
    {
       for( i = 0; i < M; i++ )
       {
-         a0 = A + LINDXA[i]; u0 = U0 + LINDXAU[i];
+         a0 = A + (size_t)(LINDXA[i]); u0 = U0 + (size_t)(LINDXAU[i]);
          for( j = 0; j < nr; j++, a0 += LDA, u0 += LDU ) { *a0 = *u0; }
       }
    }
