+++ src/pauxil/HPL_dlaswp06T.c
@@ -1,6 +1,6 @@
 /* 
  * -- High Performance Computing Linpack Benchmark (HPL)                
- *    HPL - 1.0a - January 20, 2004                          
+ *    HPL - 1.0b - December 15, 2004                          
  *    Antoine P. Petitet                                                
  *    University of Tennessee, Knoxville                                
  *    Innovative Computing Laboratories                                 
@@ -143,7 +143,8 @@ void HPL_dlaswp06T
    {
       for( i = 0; i < M; i++ )
       {
-         a0 = A + LINDXA[i]; u0 = U0 + i * LDU;
+         a0 = A  + (size_t)(LINDXA[i]);
+         u0 = U0 + (size_t)(i) * (size_t)(LDU);
 
          r = *a0; *a0 = u0[ 0]; u0[ 0] = r; a0 += LDA;
 #if ( HPL_LASWP06T_DEPTH >  1 )
@@ -194,7 +195,8 @@ void HPL_dlaswp06T
    {
       for( i = 0; i < M; i++ )
       {
-         a0 = A + LINDXA[i]; u0 = U0 + i * LDU;
+         a0 = A  + (size_t)(LINDXA[i]);
+         u0 = U0 + (size_t)(i) * (size_t)(LDU);
          for( j = 0; j < nr; j++, a0 += LDA )
          { r = *a0; *a0 = u0[j]; u0[j] = r; }
       }
